Add tests for DatastoreStorageFactory.Builder.
